+++ crates/stackable-versioned/CHANGELOG.md
@@ -9,7 +9,8 @@ All notable changes to this project will be documented in this file.
 - Add `changed()` action and support for the `ty` parameter in the `renamed()`
   action ([#844]).
 - Pass through container and item attributes (including doc-comments). Add
-  attribute for version specific docs. ([#847])
+  attribute for version specific docs ([#847]).
+- Forward container visibility to generated modules ([#850]).
 
 ### Fixed
 
@@ -19,6 +20,7 @@ All notable changes to this project will be documented in this file.
 [#842]: https://github.com/stackabletech/operator-rs/pull/842
 [#844]: https://github.com/stackabletech/operator-rs/pull/844
 [#847]: https://github.com/stackabletech/operator-rs/pull/847
+[#850]: https://github.com/stackabletech/operator-rs/pull/850
 
 ## [0.1.1] - 2024-07-10